feat: Cloud Spanner support (#8730)
* working on Cloud Spanner driver implementation

* working on DDL synchronization

* working on DDL synchronization

* fixed failing test

* working on VIEW implementation

* fixed query parameters

* lint

* added transaction support;
added streaming support;

* fixed column types

* fixes after merge

* prettier

* added support for generated columns

* added escaping for distinct alias

* working on generated columns

* changed failing test

* updated tests for Spanner;
bugfixes;

* updated tests for Spanner;
bugfixes;

* updated tests for Spanner;
bugfixes;

* fixed failing test

* fixed failing test

* fixing failing tests

* fixing failing tests

* fixing failing tests

* added support for typeorm-generated uuid;
fixed caching;

* fixing failing tests

* fixing failing tests

* fixing failing tests

* fixing failing tests

* fixing failing tests

* fixing failing tests

* debugging failing test

* debugging failing test

* fixed bug in @PrimaryColumn decorator

* fixed failing tests

* fixed VIEW functionality;
fixed failing tests;

* updated docs

+++ test/github-issues/85/issue-85.ts
@@ -24,6 +24,9 @@ describe("github issues > #85 - Column option insert: false, update: false", () 
     it("should ignore value of non-inserted column", () =>
         Promise.all(
             connections.map(async (connection) => {
+                // Skip because test relies on DEFAULT values and Spanner does not support it
+                if (connection.driver.options.type === "spanner") return
+
                 const doc1 = new Document()
                 doc1.id = 1
                 doc1.version = 42
@@ -37,6 +40,9 @@ describe("github issues > #85 - Column option insert: false, update: false", () 
     it("should be able to create an entity with column entirely missing", () =>
         Promise.all(
             connections.map(async (connection) => {
+                // Skip because test relies on DEFAULT values and Spanner does not support it
+                if (connection.driver.options.type === "spanner") return
+
                 // We delete the non-inserted column entirely, so that any use of it will throw an error.
                 let queryRunner = connection.createQueryRunner()
                 await queryRunner.dropColumn("document", "permission")
